Use Travis Buddy to comment on PRs that result in a failed CI build
Travis Buddy automatically comments on every pull request that results in a failed Travis CI build. This will cause the author of the pull request to get a notification. The idea is that the pull request author will then resolve the issue that caused the failed Travis CI build, without any intervention from a maintainer of the repository being necessary.

https://www.travisbuddy.com